apply middleware
allow whoami to anyone but restrict others

+++ app/Telegram/Middleware/OnlyDavidMiddlware.php
@@ -4,14 +4,11 @@ namespace App\Telegram\Middleware;
 
 use SergiX44\Nutgram\Nutgram;
 
-// Is this being used automatically globally?
-
-
 class OnlyDavidMiddlware
 {
     public function __invoke(Nutgram $bot, $next): void
     {
-        $davidUserId = config('nutgram.david');
+        $davidUserId = (int) config('nutgram.david');
 
         if ($bot->user()?->id !== $davidUserId) {
             $bot->sendMessage('Sorry, you are not authorized to use this bot.');
